+++ Renderer.Direct3D12/Shaders/Raytrace/Hit/SphereHitGroup.hlsl
@@ -1,8 +1,8 @@
 struct SphereHitGroupParameters
 {
-    half3 WorldPosition;
-    half Size;
-    half3 Colour;
-    half3 EmissionColour;
-    half EmissionStrength;
+    float3 WorldPosition;
+    float Size;
+    float3 Colour;
+    float3 EmissionColour;
+    float EmissionStrength;
 };
